% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circleArea.R
\name{circleArea}
\alias{circleArea}
\title{Calculate the area of a circle, given its radius}
\usage{
circleArea(r)
}
\arguments{
\item{r}{Radius of a circle}
}
\value{
Area of a circle
}
\description{
Calculate the area of a circle, given its radius
}
\examples{
circleArea(1:5)

}
