% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFirstPkg-package.r
\docType{data}
\name{newIris}
\alias{newIris}
\title{A cutdown iris dataset}
\format{A data frame with 100 rows and 5 columns}
\usage{
newIris
}
\description{
A cutdown iris dataset
}
\keyword{datasets}
